+++ SourceCode/App/Validators/ModuleValidator.cs
@@ -12,19 +12,19 @@ namespace ModulesRegistry.Validators
                 .NotEmpty()
                 .MinimumLength(3)
                 .MaximumLength(50)
-                .IsOrdinaryText(localizer)
-                .NameIsCapitalizedCorrectly(localizer)
+                .MustBeOrdinaryText(localizer)
+                .MustBeCapitalizedCorrectly(localizer)
                 .WithName(n => localizer[nameof(n.FullName)]);
             RuleFor(m => m.ScaleId)
                 .MustBeSelected(localizer)
                 .WithName(n => localizer[nameof(n.Scale)]);
             RuleFor(m => m.ConfigurationLabel)
                 .MaximumLength(10)
-                .IsOrdinaryText(localizer)
+                .MustBeOrdinaryText(localizer)
                 .WithName(n => localizer[nameof(n.ConfigurationLabel)]);
             RuleFor(m => m.PackageLabel)
                 .MaximumLength(10)
-                .IsOrdinaryText(localizer)
+                .MustBeOrdinaryText(localizer)
                 .WithName(n => localizer[nameof(n.PackageLabel)]);
             RuleFor(m => m.FremoNumber)
                 .InclusiveBetween(1, 9999).When(m => m.FremoNumber.HasValue)
@@ -34,7 +34,7 @@ namespace ModulesRegistry.Validators
                 .WithName(n => localizer[nameof(n.Standard)]);
             RuleFor(m => m.Theme)
                 .MaximumLength(50)
-                .IsOrdinaryText(localizer)
+                .MustBeOrdinaryText(localizer)
                 .WithName(n => localizer[nameof(n.Theme)]);
             RuleFor(m => m.RepresentsFromYear)
                 .MustBeValidYear(localizer)
@@ -71,7 +71,7 @@ namespace ModulesRegistry.Validators
               .WithName(n => localizer[nameof(n.SignalFeature)]);
             RuleFor(m => m.Note)
                 .MaximumLength(50)
-                .IsOrdinaryText(localizer)
+                .MustBeOrdinaryText(localizer)
                 .WithName(n => localizer[nameof(n.Note)]);
         }
     }