+++ SourceCode/App/Validators/StationValidator.cs
@@ -12,15 +12,15 @@ namespace ModulesRegistry.Validators
                 .NotEmpty()
                 .MinimumLength(3)
                 .MaximumLength(50)
-                .IsOrdinaryText(localizer)
-                .NameIsCapitalizedCorrectly(localizer)
+                .MustBeOrdinaryText(localizer)
+                .MustBeCapitalizedCorrectly(localizer)
                 .WithName(n => localizer[nameof(n.FullName)]);
             RuleFor(m => m.Signature)
                 .NotEmpty()
                 .MinimumLength(1)
                 .MaximumLength(5)
-                .IsOrdinaryText(localizer)
-                .NameIsCapitalizedCorrectly(localizer)
+                .MustBeOrdinaryText(localizer)
+                .MustBeCapitalizedCorrectly(localizer)
                 .WithName(n => localizer[nameof(n.Signature)]);
         }
     }