+++ SourceCode/App/Validators/ValidatorsExtensions.cs
@@ -1,12 +1,13 @@
 ﻿using FluentValidation;
 using Microsoft.Extensions.Localization;
 using System;
+using ModulesRegistry.Services.Extensions;
 
 namespace ModulesRegistry.Validators
 {
     public static class ValidatorsExtensions
     {
-        public static IRuleBuilderOptions<T, string> NameIsCapitalizedCorrectly<T>(this IRuleBuilder<T, string> builder, IStringLocalizer localizer) =>
+        public static IRuleBuilderOptions<T, string> MustBeCapitalizedCorrectly<T>(this IRuleBuilder<T, string> builder, IStringLocalizer localizer) =>
             builder.Must(value => value.IsNameCapitalizedCorrectly()).WithMessage($"\"{{PropertyName}}\" {localizer["MustBeginWithCapitalLetter"]}");
 
         private static bool IsNameCapitalizedCorrectly(this string? name)
@@ -21,8 +22,10 @@ namespace ModulesRegistry.Validators
             }
             return true;
         }
+        public static IRuleBuilderOptions<T, string?> MustBeOrdinaryTextOrNull<T>(this IRuleBuilder<T, string?> builder, IStringLocalizer localizer) =>
+           builder.Must(value => value.IsText()).WithMessage($"\"{{PropertyName}}\" {localizer["MayOnlyContainOrdinaryText"]}");
 
-        public static IRuleBuilderOptions<T, string> IsOrdinaryText<T>(this IRuleBuilder<T, string> builder, IStringLocalizer localizer) =>
+        public static IRuleBuilderOptions<T, string> MustBeOrdinaryText<T>(this IRuleBuilder<T, string> builder, IStringLocalizer localizer) =>
             builder.Must(value => value.IsText()).WithMessage($"\"{{PropertyName}}\" {localizer["MayOnlyContainOrdinaryText"]}");
         private static bool IsText(this string? text)
         {
@@ -39,7 +42,7 @@ namespace ModulesRegistry.Validators
 
         private static bool IsInRange(this char c, int firstCodePoint, int lastCodePoint) =>
             c >= firstCodePoint && c <= lastCodePoint;
-        private static bool IsPermittedPunctuationOrSymbol(this char c) => " (),.-&#±°²³/«»£€".Contains(c);
+        private static bool IsPermittedPunctuationOrSymbol(this char c) => " (),.-+*!?%&§#±°²³/«»£€´".Contains(c);
         private static bool IsDigit(this char c) => c >= 0x0030 && c <= 0x0039;
         private static bool IsLatinChar(this char c) =>
             c.IsInRange(0x0061, 0x007A) ||
@@ -65,5 +68,11 @@ namespace ModulesRegistry.Validators
         private static bool IsValidHour(this short? hour) => hour is null || (hour >= MinHour && hour <= MaxHour);
         const short MinHour = 0;
         const short MaxHour = 23;
+
+        public static IRuleBuilderOptions<T, string?> MustBeColor<T>(this IRuleBuilder<T, string?> builder, IStringLocalizer localizer) =>
+             builder.Must(value => value.IsHexColorOrNull()).WithMessage($"\"{{PropertyName}}\" {string.Format(localizer["MustBeAColor"].Value, MinHour, MaxHour)}");
+
+        private static bool IsHexColorOrNull(this string? value) =>
+            value is null ? true : value.IsHexColor();
     }
 }