+++ SourceCode/Services/Extensions/StringExtensions.cs
@@ -20,14 +20,11 @@ namespace ModulesRegistry.Services.Extensions
         public static string? Max(this string? me, int max) => 
             me is null ? null : me.Length < max ? me : me[0..max];
 
-        public static bool IsEmailAddress([NotNullWhen(true)] this string? me) => 
-            me.HasValue() && Regex.IsMatch(me, @"^\w+([-+.']\w+)*@\w+([-.]\w+)*\.\w+([-.]\w+)*$");
 
        public static string[] Items(this string? value, char separator = ';') =>
             string.IsNullOrWhiteSpace(value) ? Array.Empty<string>() :
             value.Trim().Split(separator);
 
-        #region Special string conversions
 
         public static string HtmlFromMarkdown(this string? markdown)
         {
@@ -42,8 +39,33 @@ namespace ModulesRegistry.Services.Extensions
             }
         }
 
+        public static bool IsEmailAddress([NotNullWhen(true)] this string? me) => 
+            me.HasValue() && Regex.IsMatch(me, @"^\w+([-+.']\w+)*@\w+([-.]\w+)*\.\w+([-.]\w+)*$");
+
         public static bool IsPermittedFileExtension(this string? it) =>
             it is not null && DocumentService.PermittedFileExtenstions.Contains(it.ToLowerInvariant());
+
+        #region Color 
+        public static string? TextColor(this string? backColor)
+        {
+            if (backColor.IsHexColor())
+            {
+                var r = int.Parse(backColor.Substring(1, 2), System.Globalization.NumberStyles.HexNumber);
+                var g = int.Parse(backColor.Substring(3, 2), System.Globalization.NumberStyles.HexNumber);
+                var b = int.Parse(backColor.Substring(5, 2), System.Globalization.NumberStyles.HexNumber);
+                var yiq = ((r * 299) + (g * 587) + (b * 114)) / 1000;
+                return (yiq >= 128) ? "#000000" : "#FFFFFF";
+            }
+            return null;
+        }
+
+        private static string HexColorRegEx => "^#([A-Fa-f0-9]{6})$";
+        public  static bool IsHexColor([NotNullWhen(true)] this string? maybeColor) =>
+            !string.IsNullOrWhiteSpace(maybeColor) && Regex.IsMatch(maybeColor, HexColorRegEx);
+
+        #endregion
+
+        #region GUID string
 
         public static Guid? AsGuid(this string? me)
         {